# AVX2 optimization of group lookup.