@greg7mdp 2022-09-13
Wow, this is great, I really appreciate it @bashimao ! I did not investigate the AVX2 code, but the change looks very safe as it is conditioned with by a compilation flag. Did you see better performance when using AVX2 versus SSE2?