@greg7mdp 2022-09-15
@bashimao I did some performance testing and unfortunately I cannot see any performance improvement with the avx2 implementation. Maybe it would show only when the hash map is close to full (just before resizing).  I'll do some more testing, but I'd like to know what you saw in your own testing.