@bashimao 2022-09-18
Hi @bashimao , sorry for the delay. I am currently travelling overseas. I get back to you later this week with some benchmark results.