@bashimao 2022-09-28
Well, that is a now a big question. Very hard to answer. I admit, I may have been chasing a ghost here. I had a use-case, and there it was faster. But marginally. And here you can see why:

![image](https://user-images.githubusercontent.com/7855770/192684777-84d13f58-71db-4f60-b3b7-f93159d8936e.png)

This is is mostly your random insertion benchmark, but for reproductivity reasons this one only benchmarks the non-parallel `flat_map`. So only one core is used. The benchmarks were run on an `AMD EPYC 7742 64-Core` processor. All results displayed represent the median-value from 5 runs. So this graphic is pretty stable.

As can be seen, we can achieve quite a decent performance without SSE already. SSE is the quickest, but not always. AVX2 occasionally overtakes, but only if you do not keep max occupancy at 87.5%. The reason being, that the AVX2 implementation reads 32 slots per group. If we limit to 7/8 (87.5%), each group we read contains 4 blanks on average. So the likelihood that we need to read another group is low. But the same is true if we use SSE where we read 16 slots per group, with an average blank rate of 2. So no real gain for the majority of reads. I think the amount of bytes read, doesn't really matter, because the cache line size of this CPU is anyway 64 byte. However, some of the AVX2 instructions have a slightly higher latency than their SSE counterparts, which I assume is the reason why it is slower. IMHO, that has to do with the larger register file in AVX2. Anyhow, the same AVX instructions will either be as fast as their SSE counterparts or a little bit slower. 

As you can see, the AVX2 implementation has the potential to overtake the SSE version. But only if we reduce to avg. 3 blanks per group, or avg. 2 blanks per group. As can be seen by the SSE implementation, having 2 blanks per group is simply enough. After all, we only need to find a single blank before we can stop searching.

However, as you may see, the side-by-side comparison breaks down there. I need to increase the max occupancy of the hash table to achieve a lower blanks-per-group-ratio. In a way, that is nice, because we are making more efficient use of our memory cells with the AVX2 version. But because that also means that we expand the table at different times, the comparison breaks down. At certain sizes, the AVX2 implementation is certainly faster. But that it is incidental because the table has a different fill state at these times.

I think merging this is fine (with some minor adjustments), but I would by default still select SSE. It is at least always faster than the non-optimized version.

2 things that I would want to explore next:
- This was measured on an AMD system. AVX on Intel CPUs is known to occasionally behave different. Should definitely check.
- Given the current findings, I doubt that switching to AVX-512 would yield a much different outcome. However, while going through the AVX-512 instruction set, I found a couple of new commands, that provide back-ports for SSE and AVX registers. In particular one of them would allow getting rid of the `movemask` instruction. That could increase the throughput of both the SSE and AVX implementations. And with this, this merge request would at least add a concrete improvement that improves the speed all users. However, I would need to find a GPU that supports the right AVX-512 command subset to test first. That may take a couple of days.