@bashimao 2022-10-06
Actually, This could be the root cause of the issue. With 64 byte, any unaligned memory access will require hitting 2 cache lines. With a 32 byte lookup like SSE, you may - more often than not - get by with reading only a single cache line. I need to investigate. Maybe it is possible to align the group lookup.

https://lemire.me/blog/2022/06/06/data-structure-size-and-cache-line-accesses/