@greg7mdp 2022-10-06
Hey @bashimao , thank for following up with this and the hard work. I think the group lookup starts at the index determined by the hash, so I'm not sure how it can be aligned, and you are right that maybe the cost of accessing more memory for every lookup is higher than the benefit for the rare occasions where the match is not found in the first 16 slots. I'm really not very knowledgeable on SSE/AVX2 so I'm afraid I don't have any specific input, but kudos for the work and good luck finding a faster implementation.