@greg7mdp 2025-04-06
> @gdb4713 @greg7mdp I am still alive. I didn't push this in the past year. Unfortunately, the product I was working on moved away from phmap recently. I think the PR code needs a little love to make it compatible with the current repo. Do you still want this change?

I would not have an issue merging the AVX2 code, so people can use it if desired, but it would have to be a manual choice, not an automatic selection at compile time. This, because I wouldn't want people recompiling with a new version of phmap have code that works on a reduced set of processors. So maybe remove the setting of `PHMAP_HAVE_AVX2` in `phmap_config.h`.

And yes, probably a little work to make the code compatible with the current repo (because of [this](https://github.com/greg7mdp/parallel-hashmap/commit/4817a6d3b8407063cf0328eb92dbb27ee2f55528) change I think).